+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# unreleased
+# 2.93.0
 * Allow PayPal payment ID and payer ID to be passed during transaction create
 * Add `travel_flight` support to industry-specific data
 * Add `fraud_service_provider` field to `risk_data`